Valida que nao possa ter dois pagamentos com mesmo ID

+++ src/main/java/br/com/lunacom/tools/repository/PagamentoRepository.java
@@ -1,9 +1,8 @@
 package br.com.lunacom.tools.repository;
 
 import br.com.lunacom.tools.domain.entity.PagamentoEntity;
-import br.com.lunacom.tools.domain.response.PagamentoResponse;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PagamentoRepository extends JpaRepository<PagamentoEntity, Integer> { }
+public interface PagamentoRepository extends JpaRepository<PagamentoEntity, Long> { }
